# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.PortfolioDetail:
    properties:
      commentCount:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      latestVersion:
        $ref: '#/definitions/handlers.VersionDTO'
      likeCount:
        type: integer
      likedByMe:
        type: boolean
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/handlers.UserPublic'
      versionCount:
        type: integer
      versions:
        items:
          $ref: '#/definitions/handlers.VersionDTO'
        type: array
    type: object
  handlers.PortfolioSummary:
    properties:
      commentCount:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      latestVersion:
        $ref: '#/definitions/handlers.VersionDTO'
      likeCount:
        type: integer
      likedByMe:
        type: boolean
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/handlers.UserPublic'
      versionCount:
        type: integer
    type: object
  handlers.RoastDTO:
    properties:
      aiGenerated:
        type: boolean
      body:
        type: string
      createdAt:
        type: string
      helpful:
        type: integer
      id:
        type: integer
      portfolio:
        properties:
          id:
            type: integer
          screenshotUrl:
            type: string
          title:
            type: string
        type: object
      stars:
        type: integer
      title:
        type: string
      user:
        $ref: '#/definitions/handlers.UserPublic'
    type: object
...
  handlers.UserPublic:
    properties:
      bio:
        type: string
      githubUrl:
        type: string
      id:
        type: integer
      initials:
        type: string
      linkedinUrl:
        type: string
      location:
        type: string
      name:
        type: string
      role:
        type: string
      twitterUrl:
        type: string
      username:
        type: string
      websiteUrl:
        type: string
    type: object
  handlers.VersionDTO:
    properties:
      codeRating:
        type: number
      createdAt:
        type: string
      id:
        type: integer
      label:
        type: string
      note:
        type: string
      number:
        type: integer
      projectUrl:
        type: string
      screenshotUrl:
        type: string
      uiRating:
        type: number
      uxRating:
        type: number
    type: object
  handlers.addVersionRequest:
    properties:
      note:
        type: string
      projectUrl:
        type: string
      screenshotUrl:
        type: string
    type: object
...
  handlers.createPortfolioRequest:
    properties:
      note:
        type: string
      projectUrl:
        type: string
      screenshotUrl:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - title
    type: object
  handlers.createRoastRequest:
    properties:
      body:
        type: string
      portfolioId:
        type: integer
      stars:
        maximum: 5
        minimum: 0
        type: integer
      title:
        type: string
    required:
    - body
    - portfolioId
    - title
    type: object
...
  handlers.updateMeRequest:
    properties:
      bio:
        type: string
      githubUrl:
        type: string
      linkedinUrl:
        type: string
      location:
        type: string
      name:
        type: string
      role:
        type: string
      twitterUrl:
        type: string
      websiteUrl:
        type: string
    type: object
  handlers.userProfileResponse:
    properties:
      stats:
        $ref: '#/definitions/handlers.userStats'
      user:
        $ref: '#/definitions/handlers.UserPublic'
    type: object
  handlers.userStats:
    properties:
      portfolioCount:
        type: integer
      rank:
        type: integer
      totalLikes:
        type: integer
    type: object
info:
  contact: {}
  description: REST API powering the DevFolio frontend — auth, portfolios, versions,
    likes, comments, roasts, leaderboard, notifications and AI Studio.
  title: FolioHub API
  version: "1.0"
paths:
...
  /portfolios:
    get:
      parameters:
      - description: trending | latest | mostliked | mostversions
        in: query
        name: sort
        type: string
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - description: Match title or author name
        in: query
        name: search
        type: string
      - description: Only the authenticated user's portfolios
        in: query
        name: mine
        type: boolean
      - description: Max results (default 30)
        in: query
        name: limit
        type: integer
      - description: Offset for paging
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.PortfolioSummary'
            type: array
      summary: List portfolios (feed / explore)
      tags:
      - portfolios
    post:
      consumes:
      - application/json
      parameters:
      - description: Portfolio + first version
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createPortfolioRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.PortfolioDetail'
      security:
      - BearerAuth: []
      summary: Post a new portfolio
      tags:
      - portfolios
  /portfolios/{id}:
    get:
      parameters:
      - description: Portfolio ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PortfolioDetail'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a portfolio's detail
      tags:
      - portfolios
...
  /portfolios/{id}/versions:
    post:
      consumes:
      - application/json
      parameters:
      - description: Portfolio ID
        in: path
        name: id
        required: true
        type: integer
      - description: New version
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.addVersionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.VersionDTO'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a new version to a portfolio
      tags:
      - portfolios
  /roasts:
    get:
      parameters:
      - description: top | new
        in: query
        name: sort
        type: string
      - description: Max results (default 30)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.RoastDTO'
            type: array
      summary: List community roasts
      tags:
      - roasts
    post:
      consumes:
      - application/json
      parameters:
      - description: Roast content
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createRoastRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.RoastDTO'
      security:
      - BearerAuth: []
      summary: Post a roast
      tags:
      - roasts
...
  /users/{username}:
    get:
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.userProfileResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user's public profile
      tags:
      - users
  /users/{username}/portfolios:
    get:
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.PortfolioSummary'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List a user's portfolios
      tags:
      - users
  /users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.userProfileResponse'
      security:
      - BearerAuth: []
      summary: Current user's profile with stats
      tags:
      - users
    patch:
      consumes:
      - application/json
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserPublic'
      security:
      - BearerAuth: []
      summary: Update my profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"